# perf: Add 8-bit lookup table to Huffman decoder