@orisano 2024-11-04
Looking forward to your new version release once this PR gets merged.