@orisano 2024-11-07
@cosnicolaou Are you planning to release the next version?