@cosnicolaou 2024-11-07
done!

Thanks for all of the perf improvements!

Cheers, Cos.

On Thu, Nov 7, 2024 at 7:59 AM Nao Yonashiro ***@***.***>
wrote:

> All the PRs I was planning to send have been sent and merged.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/cosnicolaou/pbzip2/pull/45#issuecomment-2462604935>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ACCMUYE433KK6LEAQAFNVTLZ7OE5BAVCNFSM6AAAAABRDA5UDWVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDINRSGYYDIOJTGU>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>